fix: address PHPStan errors and review comments on tax extension
- Restore missing `use Illuminate\Foundation\Application` import in
  TaxServiceProvider (caused fatal PHPStan + runtime class-not-found)
- Remove stale `@var TaxRepository $repository` PHPDoc whose variable
  no longer exists after the service provider simplification
- Replace unsafe `preg_replace` with `Str::slug()` in TaxController
  export action — resolves PHPStan unsafe-function error and handles
  more edge cases (accents, dots, etc.) per reviewer suggestion
- Revert `tax_profile_id` and `tag_id` migration columns back to
  `unsignedBigInteger` to match the FK targets (`extension_tax_profiles.id`
  and `tags.id` both use Laravel's `$table->id()` = BIGINT UNSIGNED)

+++ database/migrations/extensions/2026_03_21_000001_create_tax_tables.php
@@ -30,8 +30,8 @@ return new class extends Migration {
 
         Schema::create('extension_tax_tag_links', static function (Blueprint $table): void {
             $table->id();
-            $table->unsignedInteger('tax_profile_id');
-            $table->unsignedInteger('tag_id');
+            $table->unsignedBigInteger('tax_profile_id');
+            $table->unsignedBigInteger('tag_id');
             $table->timestamps();
 
             $table->foreign('tax_profile_id')->references('id')->on('extension_tax_profiles')->onDelete('cascade');
